+++ CHANGELOG.md
@@ -10,11 +10,14 @@ All notable changes to `Slytherin` will be documented in this file.
 - `$handler->display()` in `ErrorHandlerIntegration` instead from `Application::run`
 
 ### Removed
+- `addRoute`, `getRoute`, and `getRoutes` in `Routing\Router`
+- `setCollector` in `FastRouteRouter` and `PhrouteRouter`
 - `Component` related classes
 - `ComponentCollection`
 - `Container\Container::add` method
 - `Debug\Debugger` class
 - `Debugger` related classes
+- `Dispatching` directory
 - `Middleware\Dispatcher::__invoke` method 
 - `Middleware\Dispatcher::getStack` method
 - `Middleware\FinalResponse` class